rename spare/s3 to spare/spare_s3 to avoid conflict
