+++ tests/ingestion/spare/spare_s3_test.py
@@ -8,7 +8,7 @@ import os
 import pytest
 from unittest.mock import patch
 
-import odin.ingestion.spare.s3 as spare_s3
+import odin.ingestion.spare.spare_s3 as spare_s3
 
 
 @pytest.fixture
